change parseInt to parseFloat

+++ 04-javascript-calculator/src/components/buttons/Equals.js
@@ -22,24 +22,24 @@ function EqualsButton(props) {
       while (i < expressionArray.length) {
         let sign = 1;
         if (expressionArray[i+1] === 'x' && i+2 < expressionArray.length) {
-          let first = parseInt(expressionArray[i]);
+          let first = parseFloat(expressionArray[i]);
           // Handle negative signs
           if (expressionArray[i+2] === '-') {
             i++;
             sign = -1;
           }
-          let second = parseInt(expressionArray[i+2]) * sign;
+          let second = parseFloat(expressionArray[i+2]) * sign;
           const subtotal = first * second;
           evaluatedArray.push(subtotal.toString());
           i = i+3;
         } else if (expressionArray[i+1] === '/' && i+2 < expressionArray.length) {
-          let first = parseInt(expressionArray[i]);
+          let first = parseFloat(expressionArray[i]);
           // Handle negative signs
           if (expressionArray[i+2] === '-') {
             i++;
             sign = -1;
           }
-          let second = parseInt(expressionArray[i+2]) * sign;
+          let second = parseFloat(expressionArray[i+2]) * sign;
           const subtotal = first / second;
           evaluatedArray.push(subtotal.toString());
           i = i+3;
@@ -55,19 +55,19 @@ function EqualsButton(props) {
         if (i === 0) {
           // If first number is negative
           if (evaluatedArray[i] === '-' && !isOperator(evaluatedArray[i+1])) {
-            total -= parseInt(evaluatedArray[i+1]);
+            total -= parseFloat(evaluatedArray[i+1]);
             i++;
           } else { // If first number is positive
-            total += parseInt(evaluatedArray[i]);
+            total += parseFloat(evaluatedArray[i]);
           }
         }
         if ((i === evaluatedArray.length - 1) && (evaluatedArray[i] === 'x' || evaluatedArray[i] === '/' || evaluatedArray[i] === '+' || evaluatedArray[i] === '-')) {
         i++;
         } else if (evaluatedArray[i] === '+') {
-          total += parseInt(evaluatedArray[i+1]);
+          total += parseFloat(evaluatedArray[i+1]);
           i++;
         } else if (evaluatedArray[i] === '-') {
-          total -= parseInt(evaluatedArray[i+1]);
+          total -= parseFloat(evaluatedArray[i+1]);
           i++;
         }
       }
